Rename appsignal useragent package
